% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ClusterPalette}
\alias{ClusterPalette}
\title{An acceptable cluster color palette}
\usage{
ClusterPalette(n, vcycle = c(1, 0.7), scycle = c(0.7, 1), alpha = 1)
}
\arguments{
\item{n}{How many colors to generate}

\item{vcycle, scycle}{Small vectors with cycles of saturation/value for hsv}

\item{alpha}{Opacity of the colors}
}
\description{
An acceptable cluster color palette
}
\examples{
EmbedSOM::ClusterPalette(10)
}
