% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExpressionPalette}
\alias{ExpressionPalette}
\title{Marker expression palette generator defined from ColorBrewer's RdYlBu}
\usage{
ExpressionPalette(n, alpha = 1)
}
\description{
Marker expression palette generator defined from ColorBrewer's RdYlBu
}
