% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{GQTSOM}
\alias{GQTSOM}
\title{Train a Growing Quadtree Self-Organizing Map}
\usage{
GQTSOM(data, init.dim = c(3, 3), target_codes = 100, rlen = 10,
  radius = c(sqrt(sum(init_dim^2)), 0.1), epochRadii = seq(radius[1],
  radius[2], length.out = rlen), init = FALSE, initf = Initialize_PCA,
  coords = NULL, codes = NULL, importance = NULL, distf = 2,
  nhbr.distf = 2, noMapping = F, parallel = F, threads = if
  (parallel) 0 else 1)
}
\description{
Train a Growing Quadtree Self-Organizing Map
}
