% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{GraphCoords}
\alias{GraphCoords}
\title{Add Kamada-Kawai-generated embedding coordinates to the map}
\usage{
GraphCoords(
  dim = NULL,
  dist.method = NULL,
  distFn = function(x) x,
  layoutFn = igraph::layout_with_kk
)
}
\arguments{
\item{dim}{Dimension of the result (passed to \code{layoutFn})}

\item{dist.method}{The method to compute distances, passed to \code{\link[stats:dist]{stats::dist()}} as parameter \code{method}}

\item{distFn}{Custom transformation function of the distance matrix}

\item{layoutFn}{iGraph-compatible graph layouting function (default \link[igraph:layout_with_kk]{igraph::layout_with_kk})}
}
\value{
a function that transforms the map, usable as \code{coordsFn} parameter
}
\description{
This uses a complete graph on the map codebook, which brings overcrowding
problems. It is therefore useful to transform the distances for avoiding that
(e.g. by exponentiating them slightly).
}
