% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{Initialize_PCA}
\alias{Initialize_PCA}
\title{Create a grid from first 2 PCA components}
\usage{
Initialize_PCA(data, xdim, ydim, zdim)
}
\arguments{
\item{data}{matrix in which each row represents a point}

\item{xdim}{x dimension of the grid}

\item{xdim}{y dimension of the grid}
}
\value{
array containing the selected selected rows
}
\description{
Create a grid from first 2 PCA components
}
