% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{MapDataToCodes}
\alias{MapDataToCodes}
\title{Assign nearest node to each datapoint}
\usage{
MapDataToCodes(
  codes,
  data,
  distf = 2,
  parallel = F,
  threads = if (parallel) 0 else 1
)
}
\arguments{
\item{codes}{matrix with nodes of the SOM}

\item{data}{datapoints to assign}

\item{distf}{Distance function (1=manhattan, 2=euclidean, 3=chebyshev, 4=cosine)}

\item{threads, parallel}{Use parallel computation (see \code{\link[=SOM]{SOM()}})}
}
\value{
array with nearest node id for each datapoint
}
\description{
Assign nearest node to each datapoint
}
