% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{MSTCoords}
\alias{MSTCoords}
\title{Add MST-style embedding coordinates to the map}
\usage{
MSTCoords(
  dim = NULL,
  dist.method = NULL,
  distFn = function(x) x,
  layoutFn = igraph::layout_with_kk
)
}
\arguments{
\item{dim}{Dimension of the result (passed to layoutFn)}

\item{dist.method}{The method to compute distances, passed to \code{\link[stats:dist]{stats::dist()}} as parameter \code{method}}

\item{distFn}{Custom transformation function of the distance matrix}

\item{layoutFn}{iGraph-compatible graph layouting function (default \code{\link[igraph:layout_with_kk]{igraph::layout_with_kk()}})}
}
\value{
a function that transforms the map, usable as \code{coordsFn} parameter
}
\description{
Add MST-style embedding coordinates to the map
}
