% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NormalizeColor}
\alias{NormalizeColor}
\title{Helper for computing colors for embedding plots}
\usage{
NormalizeColor(data, low = 0, high = 1, pow = 0)
}
\description{
Helper for computing colors for embedding plots
}
