% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{UMAPCoords}
\alias{UMAPCoords}
\title{Add UMAP-based coordinates to a map}
\usage{
UMAPCoords(dim = NULL, UMAPFn = NULL)
}
\arguments{
\item{dim}{Dimension of the result (passed to \code{UMAPFn} as \code{n_components})}

\item{UMAPFn}{UMAP function to run (default \link[umap:umap]{umap::umap} configured by \link[umap:umap.defaults]{umap::umap.defaults})}
}
\value{
a function that transforms the map, usable as \code{coordsFn} parameter
}
\description{
Add UMAP-based coordinates to a map
}
