% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{UMatrixCoords}
\alias{UMatrixCoords}
\title{Add U-Matrix-optimized embedding coordinates to the map}
\usage{
UMatrixCoords(dim = NULL, dist.method = NULL, distFn = function(x) x,
  layoutFn = igraph::layout_with_kk)
}
\arguments{
\item{dim}{Dimension of the result (passed to layoutFn)}

\item{dist.method}{The method to compute distances, passed to 'stats::dist' as parameter 'method'}

\item{distFn}{Custom transformation function of the distance matrix}

\item{layoutFn}{iGraph-compatible graph layouting function (default 'igraph::layout_with_kk')}
}
\value{
a function that transforms the map, usable as 'coordsFn' parameter
}
\description{
The map must already contain a SOM grid with corresponding 'xdim','ydim' (possibly 'zdim')
}
