% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{EmbedSOM}
\alias{EmbedSOM}
\title{Process the cells with SOM into a nice embedding}
\usage{
EmbedSOM(fsom = NULL, perplexity = NULL, neighbors = NULL,
  adjust = NULL, data = NULL, map = NULL, importance = NULL)
}
\arguments{
\item{fsom}{FlowSom object with a built SOM}

\item{perplexity}{How many SOM neighbors to take seriously}

\item{neighbors}{How many SOM neighbors to take into the whole computation}

\item{data}{Data matrix with points that optionally overrides the one from `fsom$data`}

\item{map}{Map object in FlowSOM format, to optionally override `fsom$map`}
}
\value{
matrix with 2-D coordinates of the embedded cels
}
\description{
Process the cells with SOM into a nice embedding
}
