% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExpressionGradient}
\alias{ExpressionGradient}
\title{The ggplot2 scale gradient from ExpressionPalette.}
\usage{
ExpressionGradient(...)
}
\arguments{
\item{...}{Arguments passed to ggplot2::scale_color_gradientn}
}
\description{
The ggplot2 scale gradient from ExpressionPalette.
}
