% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{kMeansMap}
\alias{kMeansMap}
\title{Create a map from k-Means clusters}
\usage{
kMeansMap(data, k, coordsFn, ...)
}
\arguments{
\item{data}{Input data matrix, with individual data points in rows}

\item{k}{How many points to sample}
}
\value{
map object (without the grid, if coordsFn was not specified)
}
\description{
May give better results than 'RandomMap' on data where random sampling
is complicated.
}
\examples{
d <- iris[,1:4]
EmbedSOM::PlotEmbed(
  EmbedSOM::EmbedSOM(
    data = d,
    map = EmbedSOM::kMeansMap(d, 10, EmbedSOM::GraphCoords())),
  pch=19, clust=iris[,5]
)
}
