% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{kNNCoords}
\alias{kNNCoords}
\title{Add KNN-topology-based embedding coordinates to the map}
\usage{
kNNCoords(
  k = 4,
  dim = NULL,
  distFn = function(x) x,
  layoutFn = igraph::layout_with_kk
)
}
\arguments{
\item{k}{Size of the neighborhoods (default 4)}

\item{dim}{Dimension of the result (passed to \code{layoutFn})}

\item{distFn}{Custom transformation function of the distance matrix}

\item{layoutFn}{iGraph-compatible graph layouting function (default \link[igraph:layout_with_kk]{igraph::layout_with_kk})}
}
\value{
a function that transforms the map, usable as \code{coordsFn} parameter
}
\description{
Internally, this uses \code{\link[FNN:get.knn]{FNN::get.knn()}} to compute the k-neighborhoods. That
function only supports Euclidean metric, therefore \code{kNNCoords} throws a warning whenever
a different metric is used.
}
