% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NormalizeColor}
\alias{NormalizeColor}
\title{Helper for computing colors for embedding plots}
\usage{
NormalizeColor(data, low = 0, high = 1, pow = 0)
}
\arguments{
\item{data}{Vector of scalar values to normalize between 0 and 1}

\item{low, high}{Quantiles to be clamped to 0 and 1 (try low=0.05, high=0.95)}

\item{pow}{The scaled data are transformed to data^(2^pow). If set to 0,
nothing happens. Positive values highlight differences in the
data closer to 1, negative values highlight differences closer to 0.}
}
\description{
Helper for computing colors for embedding plots
}
\examples{
EmbedSOM::NormalizeColor(c(1,100,500))
}
