% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PlotData}
\alias{PlotData}
\title{Export a data frame for plotting with marker intensities and density.}
\usage{
PlotData(embed, fsom, data = fsom$data, cols, names, normalize = cols,
  qlimit = 0, qlow = qlimit, qhigh = 1 - qlimit, pow = 0,
  vf = PlotId, density = "Density", densBins = 256,
  densLimit = NULL, fdens = sqrt)
}
\arguments{
\item{embed, fsom, data, cols}{The embedding data, columns to select}

\item{names}{Column names for output}

\item{normalize}{List of columns to normalize using NormalizeColor, default all}

\item{qlimit, qlow, qhigh, pow}{Parameters for the normalization}

\item{vf}{Custom value-transforming function}

\item{density}{Name of the density column}

\item{densBins}{Number of bins for density calculation}

\item{densLimit}{Upper limit of density (prevents outliers)}

\item{fdens}{Density-transforming function; default sqrt}
}
\description{
Export a data frame for plotting with marker intensities and density.
}
