% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PlotEmbed}
\alias{PlotEmbed}
\title{Helper function for plotting the embedding}
\usage{
PlotEmbed(embed, value = 0, red = 0, green = 0, blue = 0,
  fr = PlotId, fg = PlotId, fb = PlotId, fv = PlotId, powr = 0,
  powg = 0, powb = 0, powv = 0, clust = NULL, nclust = 0,
  nbin = 256, maxDens = NULL, fdens = sqrt, limit = 0.01,
  pch = ".", alpha = NULL, cex = 1, exlim = 1, fsom, data, map,
  xdim, ydim, col)
}
\arguments{
\item{embed}{The embedding from EmbedSOM}

\item{value}{The column of data to use for plotting the value}

\item{red, green, blue}{The same for RGB components}

\item{fv, fr, fg, fb}{Functions to transform the values before they are normalized}

\item{powv, powr, powg, powb}{Adjustments of the value plotting}

\item{clust, nclust}{integer cluster label column/data, optional cluster number}

\item{limit}{Low/high offset for NormalizeColor}

\item{pch, cex}{Parameters for point plots}

\item{alpha}{Default alpha value}

\item{exlim}{Extra border around the embedding, default 1}

\item{fsom}{FlowSOM object}

\item{data}{Data matrix, taken from fsom parameter by default}

\item{xdim, ydim}{Sizes of the SOM, taken from fsom by default}

\item{col}{Different coloring, if supplied}
}
\description{
Takes the 'embed' object which is the output of EmbedSOM, together with a
multitude of arguments that override how the plotting is done.
}
