% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PlotGG}
\alias{PlotGG}
\title{Wrap PlotData result in ggplot object.}
\usage{
PlotGG(embed, ...)
}
\arguments{
\item{embed}{Embedding data}

\item{...}{Extra arguments passed to PlotData}
}
\description{
This creates a ggplot2 object for plotting.
}
\examples{
\dontrun{
# Default plotting:
PlotGG(e) + ggplot2::geom_point()

# Slight point style modification is recommended:
PlotGG(e, fsom=fs + geom_point(aes(color=yourColName), alpha=.3, size=.3)
}
}
