% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PlotGG}
\alias{PlotGG}
\title{Wrap PlotData result in ggplot object.}
\usage{
PlotGG(embed, ...)
}
\arguments{
\item{embed}{Embedding data}

\item{...}{Extra arguments passed to \code{\link[=PlotData]{PlotData()}}}
}
\description{
This creates a ggplot2 object for plotting.
}
\examples{
library(EmbedSOM)
library(ggplot2)

# simulate a simple dataset
e <- cbind(rnorm(10000),rnorm(10000))

PlotGG(e, data=data.frame(Expr=runif(10000))) +
  geom_point(aes_string(color="Expr"))
}
