% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{SOM}
\alias{SOM}
\title{Build a self-organizing map}
\usage{
SOM(data, xdim = 10, ydim = 10, zdim = NULL, batch = F,
  rlen = 10, alphaA = c(0.05, 0.01),
  radiusA = stats::quantile(nhbrdist, 0.67) * c(1, 0), alphaB = alphaA
  * c(-negAlpha, -0.1 * negAlpha), radiusB = negRadius * radiusA,
  negRadius = 1.33, negAlpha = 0.1, epochRadii = seq(radiusA[1],
  radiusA[2], length.out = rlen), init = FALSE, initf = Initialize_PCA,
  distf = 2, codes = NULL, importance = NULL, coordsFn = NULL,
  nhbr.method = "maximum", noMapping = F, parallel = F,
  threads = if (parallel) 0 else 1)
}
\arguments{
\item{data}{Matrix containing the training data}

\item{xdim}{Width of the grid}

\item{ydim}{Hight of the grid}

\item{zdim}{Depth of the grid, causes grid to be 3D}

\item{batch}{Use batch training (default FALSE chooses online training, more compatible with FlowSOM)}

\item{rlen}{Number of training epochs; or number of times to loop over the training data in online training}

\item{alphaA}{Start and end learning rate for online learning (only for online training)}

\item{radiusA}{Start and end radius}

\item{alphaB}{Start and end learning rate for the second radius (only for online training)}

\item{radiusB}{Start and end radius (only for online training; make sure it's larger than radiusA)}

\item{negRadius}{easy way to set radiusB as a multiple of default radius
(use lower value for higher dimensions)}

\item{negAlpha}{the same for alphaB}

\item{epochRadii}{Vector of length 'rlen' with precise epoch radii (only for batch training)}

\item{init}{Initialize cluster centers in a non-random way}

\item{initf}{Use the given initialization function if init==T
(default: Initialize_PCA)}

\item{distf}{Distance function (1=manhattan, 2=euclidean, 3=chebyshev)}

\item{codes}{Cluster centers to start with}

\item{importance}{array with numeric values. Parameters will be scaled
according to importance}

\item{coordsFn}{Function to generate/transform grid coordinates (e.g. 'EmbedSOM::tSNECoords()'). If NULL (default), the grid is the canonical SOM grid.}

\item{nhbr.method}{Way of computing grid distances, passed as method= to dist() function. Default 'maximum' (square neighborhoods); use 'euclidean' for round neighborhoods.}

\item{noMapping}{If true, do not produce mapping (default F). Useful for online/streaming use.}

\item{parallel}{Parallelize batch training (has no effect on online training) by setting appropriate 'threads'. Defaults to FALSE. Use 'batch=T' for fully parallelized version, online training is not parallelizable. Passed to 'MapDataToCodes'.}

\item{threads}{Number of threads of batch training (has no effect on online training). Defaults to 0 (chooses maximum available hardware threads) if parallel=TRUE or 1 (single thread) if parallel=FALSE. Passed to 'MapDataToCodes'.}
}
\value{
A map useful for embedding ('EmbedSOM' function) or further analysis.
}
\description{
Build a self-organizing map
}
\seealso{
FlowSOM::SOM
}
