% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{tSNECoords}
\alias{tSNECoords}
\title{Add tSNE-based coordinates to a map}
\usage{
tSNECoords(dim = NULL, tSNEFn = Rtsne::Rtsne, ...)
}
\arguments{
\item{dim}{Dimension of the result (passed to \code{tSNEFn} as \code{dims})}

\item{tSNEFn}{tSNE function to run (default \link[Rtsne:Rtsne]{Rtsne::Rtsne})}

\item{...}{passed to \code{tSNEFn}}
}
\value{
a function that transforms the map, usable as \code{coordsFn} parameter
}
\description{
Add tSNE-based coordinates to a map
}
