% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{UMAPCoords}
\alias{UMAPCoords}
\title{Add UMAP-based coordinates to a map}
\usage{
UMAPCoords(dim = NULL, UMAPFn = function(x, dim) {     cfg <-
  umap::umap.defaults     cfg$n_components = dim     umap::umap(x,
  cfg)$layout })
}
\arguments{
\item{dim}{Dimension of the result (passed to UMAPFn as 'n_components')}

\item{UMAPFn}{UMAP function to run (default 'umap::umap' configured by 'umap::umap.defaults')}
}
\value{
a function that transforms the map, usable as 'coordsFn' parameter
}
\description{
Add UMAP-based coordinates to a map
}
