% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{uwotCoords}
\alias{uwotCoords}
\title{Add UMAP-based coordinates to a map, using the 'uwot' package}
\usage{
uwotCoords(dim = NULL, uwotFn = uwot::umap, ...)
}
\arguments{
\item{dim}{Dimension of the result (passed to \code{uwotFn} as \code{dims})}

\item{uwotFn}{UMAP function to run (default \link[uwot:umap]{uwot::umap})}

\item{...}{passed to \code{uwotFn}}
}
\value{
a function that transforms the map, usable as \code{coordsFn} parameter
}
\description{
Add UMAP-based coordinates to a map, using the 'uwot' package
}
